Add executor
The executor enables to execute simple programs for the first time.

It is more like a proof of concept: it is possible to execute the
following program:

fn main() {
	print("hello, world");
}

In order to execute the code you have to write it into a file, name it
e.g. "program" and then type

	ontc run ./program

in order to call the interpreter and run the program.

The executor invokes the parser in order to parse the code and build
the AST. The parser is no longer a standalone binary anymore.
In addition, it requires a file and no longer accepts input from stdin
from now on.

After parsing, the executor finds the main method and processes the
code. It supports one built-in function called "print" which accepts
one string as an argument and prints it out to stdout.

This proof of concept is a starting point for the development of the
interpreter.