Add support for more binary operators to AST
ast_new_binop() can now handle more binary operators (including logical
and bitwise operators) so that a majority of expressions is now
supported by the AST.

ast_new_binop_s() allows to easily handle operators with more than one
character (strings) as it maps the operators to a special range in the
char type (used by ASCII control chars which are not used by this
component in any way) so that these operators fit in the char type.
After that, the expressions are processed further by ast_new_binop().