Add AST nodes for classes and sequences
Classes can now be represented in the AST so that it is completed more
and more.

Imperative blocks are now called compound statements. They are
represented as ANT_CMPD nodes and can be converted to ANT_SEQ nodes
which have the same structure as ANT_CMPD nodes. ANT_SEQ nodes are used
to distinct between blocks as function or loop bodies and nested blocks
inside functions.

ANT_ADDRESS was abbreviated to ANT_ADDR.