Restructure AST nodes
Values of nodes with the types ANT_INT, ANT_FLOAT and ANT_STR are now
stored in additional structs. The former approach relies on casting
between structs with a common sequence of types and names at the
beginning, but this approach violates the strict aliasing rule.

The new approach follows the strict aliasing rule and introduces the new
member "value" in the ast_node struct.

"value" points to a type-specific struct containing the leaf values.