Refactor and reorganize AST nodes
The second rework of the structure of the AST nodes reduces memory
usage and harmonizes the different AST node structs.

All AST nodes share a common "base" member as their first member now.
It is an ast_node_base struct and contains the node type as well as
the necessary pointers.

This restructuring obsoletes the "value" pointer. The structure now
reflects the original approach more but is implemented in a way to be
compliant to the strict aliasing rule. This is achieved by moving the
common members to the ast_node_base struct which is the first member of
each AST node struct.

Several helper macros were added to simplify the code.